backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "alembic>=1.14.1",
    "anyio>=3.7.1",
    "asyncpg>=0.28.0",
    "celery>=5.4.0",
    "fastapi>=0.103.1",
    "h11>=0.14.0",
    "httpx>=0.28.1",
    "jsonschema>=4.23.0",
    "pillow>=11.1.0",
    "pydantic>=2.4.1",
    "pydantic-core>=2.10.1",
    "pydantic-settings>=2.0.3",
    "pyspellchecker>=0.8.2",
    "pytesseract>=0.3.13",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "python-dotenv>=1.0.0",
    "python-levenshtein>=0.27.1",
    "redis>=5.2.1",
    "requests>=2.31.0",
    "sniffio>=1.3.0",
    "sqlalchemy>=2.0.20",
    "starlette>=0.27.0",
    "tenacity>=9.0.0",
    "uvicorn>=0.23.2",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "deptry>=0.23.0",
    "mypy>=1.15.0",
    "vulture>=2.14",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["src"]
